Removed recursion for Windows WMI commands
[slow]
The location of the WMI commands have been moved to the same location as the UNIX `ps` process generation. The loop to associate the process IDs and parent process IDs can then be used for all OS's.

The condition used to check if the loop should finish has also been reversed, so that it operates correctly